requirements.txt
absl-py==0.6.1
asn1crypto==0.24.0
astor==0.7.1
atomicwrites==1.2.1
attrs==18.2.0
awscli==1.16.67
boto3==1.9.57
botocore==1.12.57
cachetools==3.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
colorama==0.3.9
cryptography==2.3.1
cycler==0.10.0
Cython==0.29.1
dask==1.0.0
decorator==4.3.0
docutils==0.14
deepdiff==3.3.0
flatbuffers==1.10
funcsigs==1.0.2
future==0.17.1
gast==0.2.0
gitdb2==2.0.5
GitPython==2.1.11
glfw==1.7.0
google-api-python-client==1.7.5
google-auth==1.6.1
google-auth-httplib2==0.0.3
grpcio==1.16.1
gtimer==1.0.0b5
git+https://github.com/openai/gym.git@49cd48020f6760630a7317cb3529a22de6f12f2e#[all]
h5py==2.8.0
httplib2==0.12.0
idna==2.7
imageio==2.4.1
jmespath==0.9.3
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lockfile==0.12.2
Markdown==3.0.1
matplotlib==3.0.2
more-itertools==4.3.0
git+https://github.com/vitchyr/multiworld.git@d76b3dae2e8cbca02924f93d6cc0239c552f6408
networkx==2.2
numpy==1.15.4
pandas==0.23.4
Pillow==5.3.0
plotly==1.9.6
pluggy==0.8.0
protobuf==3.6.1
py==1.7.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycosat==0.6.3
pycparser==2.19
pygame==1.9.4
pyglet==1.3.2
pyOpenSSL==18.0.0
pyparsing==2.3.0
PySocks==1.6.8
pytest==4.0.1
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
psutil==5.4.8
ray[rllib,debug]==0.6.1
redis==3.0.1
requests==2.20.1
rsa==3.4.2
ruamel-yaml==0.15.46
s3transfer==0.1.13
scikit-image==0.14.1
scikit-learn==0.20.1
scipy==1.1.0
git+https://github.com/hartikainen/serializable.git@76516385a3a716ed4a2a9ad877e2d5cbcf18d4e6
setproctitle==1.1.10
six==1.11.0
smmap2==2.0.5
tensorboard==1.12.0
tensorflow==1.12.0
tensorflow-probability==0.5.0
termcolor==1.1.0
toolz==0.9.0
uritemplate==3.0.0
urllib3==1.23
Werkzeug==0.14.1
